feat: allow to add data with container
