+++ src/view-server/index.ts
@@ -2,3 +2,4 @@ export * from './server';
 export * from './types';
 export * from './utils';
 export * from './data-viewer';
+export * from './data-container';
